requirements.txt
keras==2.7.0
keras_visualizer==2.4
matplotlib==3.4.3
numpy==1.20.3
pandas==1.3.4
pycryptodome==3.14.1
scikit_learn==1.0.2
scipy==1.7.1
seaborn==0.11.2
tensorflow==2.7.0
tensorflow_probability==0.15.0
tqdm==4.62.3
